18-03-25

+++ src/tema7/ejercicio1/Figura2.java
@@ -3,6 +3,6 @@ package tema7.ejercicio1;
 public interface Figura2 {
     double area();
     double perimeter();
-    void escalar(double escala);
+    void escalar(int escala);
     void imprimir();
 }
